requirements.txt
aiogram==3.4.1
supabase==2.3.4
openai==1.12.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.6.1
pydantic-settings==2.2.1
PyMuPDF==1.24.0
python-docx==1.1.0
httpx==0.27.0
